mplayer: Use VA-API, multiple threads and stop xscreensaver
VA-API is better with Intel gfx cards: it will use HW decoding if
possible, and fall-back to XvMC/Xv. The setting for stopping the
XScreenSaver also works for others (e.g. gnome-screensaver-daemon).
Using multiple threads makes things smoother on multi-core systems
avoiding the need to bump up the CPU speed by spreading load to
the different cores (battery is saved in the end.)