vimrc: Do not highlight EOL spaces in cursor line
Spaces at the end of the line in insert mode are not highlighted now at the
line that contains the cursor. This is good because it prevents the matches
to appear and disappear in quick progression while writing, which can be
visually annoying. A couple of autocommands are used to enable highlighting
also the cursor line when exiting insert mode.