dotdot: Run executables and use the output as target contents
If a dot.* file is executable, it will be executed instead of symlinking
it, and its output used as the contents of the target configuration file.

Optionally, a generator may want to use the first argument passed to it,
which is the full path to the destination file, and then not producing any
output at all on stdout. Any output on stedout will overwrite whatever
contents the generator may have written to the target file.