icecc-mkenv: Add script to create IceCC compiler tarballs
The one included with IceCC wouldn't create working tarballs half of
the time for GCC, and 90% of the time they would not work for Clang.
This script does always work, and on top of that is much faster, by
caching the generated tarballs under ($XDG_CACHE_DIR), and skipping
the slow part of compressing files when the set of files being added
to the tarballs don't change.

Note that this is tested on GNU/Linux only for now.